chore: 代码调整&增加注释
